+++ README.md
@@ -8,6 +8,8 @@ data. The data is provided in a large struct you can consume in your application
 This library also contains methods to export the Unifi data in InfluxDB format,
 and this can be used as an example to base your own metrics collection methods.
 
+If more features are requested, I'll certainly consider them. Do you need to do
+more than just collect data? [Let me know](https://github.com/golift/unifi/issues/new)!
 Pull requests and feedback are welcomed!
 
 Here's a working example:
@@ -28,18 +30,25 @@ func main() {
 	// Log with log.Printf or make your own interface that accepts (msg, fmt)
 	uni.ErrorLog = log.Printf
 	uni.DebugLog = log.Printf
-	clients, err := uni.GetClients()
+
+	sites, err := uni.GetSites()
 	if err != nil {
 		log.Fatalln("Error:", err)
 	}
+	clients, err := uni.GetClients(sites)
+	if err != nil {
+		log.Fatalln("Error:", err)
+	}
+	devices, err := uni.GetDevices(sites)
+	if err != nil {
+		log.Fatalln("Error:", err)
+	}
+
 	log.Println(len(clients.UCLs), "Clients connected:")
 	for i, client := range clients.UCLs {
 		log.Println(i+1, client.ID, client.Hostname, client.IP, client.Name, client.LastSeen)
 	}
-	devices, err := uni.GetDevices()
-	if err != nil {
-		log.Fatalln("Error:", err)
-	}
+
 	log.Println(len(devices.USWs), "Unifi Switches Found")
 	log.Println(len(devices.USGs), "Unifi Gateways Found")
 
@@ -48,5 +57,4 @@ func main() {
 		log.Println(i+1, uap.Name, uap.IP)
 	}
 }
-
 ```
