+++ types_test.go
@@ -1,0 +1,36 @@
+package unifi
+
+import (
+	"encoding/json"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestFlexInt(t *testing.T) {
+	t.Parallel()
+	a := assert.New(t)
+	type testReply struct {
+		Five    FlexInt `json:"five"`
+		Seven   FlexInt `json:"seven"`
+		Auto    FlexInt `json:"auto"`
+		Channel FlexInt `json:"channel"`
+	}
+	var r testReply
+	// test unmarshalling the custom type three times with different values.
+	a.Nil(json.Unmarshal([]byte(`{"five": "5", "seven": 7, "auto": "auto"}`), &r))
+
+	// test number in string.
+	a.EqualValues(5, r.Five.Val)
+	a.EqualValues("5", r.Five.Txt)
+	// test number.
+	a.EqualValues(7, r.Seven.Val)
+	a.EqualValues("7", r.Seven.Txt)
+	// test string.
+	a.EqualValues(0, r.Auto.Val)
+	a.EqualValues("auto", r.Auto.Txt)
+	// test (error) struct.
+	a.NotNil(json.Unmarshal([]byte(`{"channel": {}}`), &r),
+		"a non-string and non-number must produce an error.")
+	a.EqualValues(0, r.Channel.Val)
+}
