+++ usg_influx.go
@@ -17,7 +17,7 @@ func (u USG) Points() ([]*influx.Point, error) {
 		"device_type":            u.Stat.O,
 		"device_oid":             u.Stat.Oid,
 		"site_id":                u.SiteID,
-		"addopted":               strconv.FormatBool(u.Adopted),
+		"adopted":                strconv.FormatBool(u.Adopted),
 		"name":                   u.Name,
 		"adopt_ip":               u.AdoptIP,
 		"adopt_url":              u.AdoptURL,
@@ -92,7 +92,7 @@ func (u USG) Points() ([]*influx.Point, error) {
 		"wan1_rx_packets":         u.Wan1.RxPackets,
 		"wan1_type":               u.Wan1.Type,
 		"wan1_speed":              u.Wan1.Speed,
-		"wan1_up":                 u.Wan1.Up.Bool,
+		"wan1_up":                 u.Wan1.Up.Val,
 		"wan1_tx_bytes":           u.Wan1.TxBytes,
 		"wan1_tx_bytes-r":         u.Wan1.TxBytesR,
 		"wan1_tx_dropped":         u.Wan1.TxDropped,
@@ -170,7 +170,7 @@ func (u USG) Points() ([]*influx.Point, error) {
 			"rx_packets":             p.RxPackets,
 			"tx_bytes":               p.TxBytes,
 			"tx_packets":             p.TxPackets,
-			"up":                     p.Up.String,
+			"up":                     p.Up.Txt,
 			"vlan":                   p.Vlan,
 			"dhcpd_ntp_1":            p.DhcpdNtp1,
 			"dhcpd_unifi_controller": p.DhcpdUnifiController,
